added README.md and refactoring
