+++ components/bl-gallery-component/component.json
@@ -21,101 +21,101 @@
       "label": "Gap",
       "type": "text",
       "defaultValue": "5px",
-      "handlerId": "gap",
-      "handlerDescription": "This is a handler to set the distance between images."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "imageHeight",
       "label": "Image Height",
       "type": "text",
       "defaultValue": "200px",
-      "handlerId": "imageHeight",
-      "handlerDescription": "This is a handler to set image height."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "borderRadius",
       "label": "Image Border Radius",
       "type": "text",
       "defaultValue": "5px",
-      "handlerId": "borderRadius",
-      "handlerDescription": "This is a handler to set image border radius."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "wrapAround",
       "label": "Wrap Around",
       "type": "checkbox",
-      "handlerId": "wrapAround",
-      "handlerDescription": "This is a handler for toggling gallery loops.  If true, when a user reaches the last image in a set, the right navigation arrow will appear and they will be to continue moving forward which will take them back to the first image in the set."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "alwaysShowNavOnTouchDevices",
       "label": "Nav On Mobile",
       "type": "checkbox",
-      "handlerId": "navOnMobile",
-      "handlerDescription": "This is a handler to enable navigation for touchscreens. If true, the left and right navigation arrows which appear on mouse hover when viewing image sets will always be visible on devices that support touch.\n"
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "showImageNumberLabel",
       "label": "Show Image Number Label",
       "type": "checkbox",
       "defaultValue": true,
-      "handlerId": "showImageNumberLabel",
-      "handlerDescription": "This is a handler for toggle image label and separator block visibility. "
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "disableScrolling",
       "label": "Disable Scrolling",
       "type": "checkbox",
-      "handlerId": "disableScrolling",
-      "handlerDescription": "This is a handler for toggle scrolling. If true, prevent the page from scrolling while Lightbox is open."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "imageLabel",
       "label": "Image Label",
       "type": "text",
       "defaultValue": "Image",
-      "handlerId": "imageLabel",
-      "handlerDescription": "This is a handler to set image label."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "separator",
       "label": "Separator",
       "type": "text",
       "defaultValue": "of",
-      "handlerId": "separator",
-      "handlerDescription": "This is a handler to set separator."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "fadeDuration",
       "label": "Fade Duration(ms)",
       "type": "number",
       "defaultValue": "600",
-      "handlerId": "fadeDuration",
-      "handlerDescription": "This is a handler to set fade duration. The time it takes for the Lightbox container and overlay to fade in and out, in milliseconds."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "imageFadeDuration",
       "label": "Image Fade Duration(ms)",
       "type": "number",
       "defaultValue": "600",
-      "handlerId": "imageFadeDuration",
-      "handlerDescription": "This is a handler to set image fade duration. The time it takes for the image to fade in once loaded, in milliseconds."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "positionFromTop",
       "label": "Position From Top(px)",
       "type": "number",
       "defaultValue": "50",
-      "handlerId": "postionFromTop",
-      "handlerDescription": "This is a handler to set the distance from top of viewport that the Lightbox container will appear, in pixels."
+      "handlerId": "",
+      "handlerDescription": ""
     },
     {
       "name": "resizeDuration",
       "label": "Resize Duration(ms)",
       "type": "number",
       "defaultValue": "700",
-      "handlerId": "resizeDuration",
-      "handlerDescription": "This is a handler to set the time it takes for the Lightbox container to animate its width and height when transition between different size images, in milliseconds."
+      "handlerId": "",
+      "handlerDescription": ""
     }
   ],
   "eventHandlers": [],
