+++ components/bl-gallery-component/src/lib/lightbox-jquery.js
