+++ components/bl-gallery-component/styles/index.less
@@ -6,7 +6,8 @@
   flex-wrap: wrap;
   justify-content: center;
   
-  &__image {
+  .image {
     max-width: 100%;
+    width: auto;
   }
 }
