+++ components/bl-gallery-component/styles/lightbox.less
